requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0

# Machine Learning & Data Science
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.25.2
scipy==1.11.4

# Database
sqlalchemy==2.0.23

# Authentication & Security
passlib==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.1.2

# Data Validation
pydantic==2.5.2

# HTTP Requests
httpx==0.25.2
requests==2.31.0

# Utilities
python-multipart==0.0.6
python-dotenv==1.0.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1